Contributes to issue #61.
I've split out sources.py into a package. general.py holds the ExtendedSource and PointSource classes. extended.py is currently only for GalaxyCluster, but may eventually containing a merging cluster class.
