+++ xga/sources/__init__.py
@@ -1,0 +1,8 @@
+#  This code is a part of XMM: Generate and Analyse (XGA), a module designed for the XMM Cluster Survey (XCS).
+#  Last modified by David J Turner (david.turner@sussex.ac.uk) 01/09/2020, 16:11. Copyright (c) David J Turner
+
+from .base import BaseSource
+from .extended import GalaxyCluster
+from .general import ExtendedSource, PointSource
+
+
